pyproject.toml
[tool.poetry]
name = "dev-cluster"
version = "0.1.0"
description = ""
authors = ["Andrei Chenchik <andrei@chenchik.me>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
pulumi = "^3.23.0"
pulumi-kubernetes = "^3.15.0"
pydantic = "^1.9.0"
bcrypt = "^3.2.0"
argon2-cffi = "^21.3.0"

[tool.poetry.dev-dependencies]
mypy = "^0.931"
black = "^21.12b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
